==> 01_move_files_to_subfolder.ipynb <==
import glob
from PyPDF2 import PdfReader
import pandas as pd
import numpy as np
import docx
from striprtf.striprtf import rtf_to_text
import sys
import os
from bs4 import BeautifulSoup
import comtypes.client #throwing error
import shutil
import win32com.client
import pytesseract
from PIL import Image
from docx.api import Document
import time
import regex as re
max_rec = 0x100000
sys.setrecursionlimit(max_rec)
pytesseract.pytesseract.tesseract_cmd = "C:\\Users\\HaoLuo\\AppData\\Local\\Programs\\Tesseract-OCR\\tesseract.exe"

st = time.time()
file_source = 'C:\\Users\\HaoLuo\\ACF-2023-0011\\ACF-2023-0011'
os.chdir(file_source)
directory = glob.glob('*')
for file in directory:
    if '.png' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Other\\'
        shutil.copy2(file_source+'\\'+file, file_destination)
    elif '.jpg' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Other\\'
        shutil.move(file_source+'\\'+file, file_destination)
    elif file[-5:]=='.docx':
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Docx\\'
        shutil.copy2(file_source+'\\'+file, file_destination)
    elif '.rtf' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Rtf\\'
        shutil.copy2(file_source+'\\'+file, file_destination)
    elif '.txt' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Txt\\'
        shutil.copy2(file_source+'\\'+file, file_destination)
    elif '.pdf' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Pdf\\'
        shutil.move(file_source+'\\'+file, file_destination)
    elif '.pptx' in file:
        file_destination = 'C:\\Users\\HaoLuo\\2023_Attachments\\Pptx\\'
        shutil.move(file_source+'\\'+file, file_destination)    
print('Different Folder: ', time.time()-st)

# Converting pptx file into PDF
def PPTtoPDF(filename, formatType = 32):
    powerpoint = comtypes.client.CreateObject("Powerpoint.Application")
#    powerpoint.Visible = 1
    in_file = os.path.abspath(filename)
    out_file = os.path.abspath(filename[0:-5]+'.pdf')
    deck = powerpoint.Presentations.Open(in_file)
    deck.SaveAs(out_file, formatType) # formatType = 32 for ppt to pdf
    deck.Close()
    powerpoint.Quit()

st = time.time()
path2 ='C:/Users/HaoLuo/2023_Attachments/Pptx' # set your own path
path3 ='C:/Users/HaoLuo/2023_Attachments/Pdf'
os.chdir(path2)
for file_name in glob.glob('*.pptx'):
    print('Converting: ', file_name, ' to PDF' )
    PPTtoPDF(file_name)
    #shutil.copy2(path2+'/'+file_name[0:-5]+'.pdf', path30+'/'+file_name[0:-5]+'.pdf')
    shutil.move(path2+'/'+file_name[0:-5]+'.pdf', path3+'/'+file_name[0:-5]+'.pdf')
print('Converting pptx to pdf into PDF folder: ', time.time()-st)
