==> 02_clean_raw_data.ipynb <==
import glob
from PyPDF2 import PdfReader
import pandas as pd
import numpy as np
import docx
from striprtf.striprtf import rtf_to_text
import sys
import os
from bs4 import BeautifulSoup
import comtypes.client #throwing error
import shutil
import win32com.client
import pytesseract
from PIL import Image
from docx.api import Document
import time
import regex as re
max_rec = 0x100000
sys.setrecursionlimit(max_rec)
pytesseract.pytesseract.tesseract_cmd = "C:\\Users\\HaoLuo\\AppData\\Local\\Programs\\Tesseract-OCR\\tesseract.exe"
st = time.time()

def scrape_html(soup):
  '''retrieve relevant content from b4s'''

  #init empty dict
  html_dict = dict()

  #put all 'b' tags into dictionary to capture meta data
  for tag in soup.find_all(name='b'):
    html_dict[tag.contents[0].split(':')[0]] = tag.next_sibling.string

  #comment is two siblings behind heading
    if soup.find(name='h2',string="General Comment") is not None:
        html_dict["Comment"] = soup.find(name='h2',string="General Comment").next_sibling.next_sibling.text
    elif soup.find(name='h2',string="Redacted Comment") is not None:
        html_dict["Comment"] = soup.find(name='h2',string="Redacted Comment").next_sibling.next_sibling.text
        
  #We only care if attachment exists. Attachment must be grabbed using the comment name, not the attachment name
  if soup.find(name='h2',string="Attachments"):
    html_dict["has_attachments"] = True
  else:
    html_dict["has_attachments"] = False

  return html_dict

def scrape_pdf(file_name):
  ''' scrape pdf '''

  reader = PdfReader(file_name)
  text = ""
  for page in reader.pages:
      text += page.extract_text() + "\n"

  return text

def scrape_rtf(filename):
    with open(filename) as infile:
        content = infile.read()
        text = rtf_to_text(content)
    return text

def scrape_txt(filename):
    with open(filename) as infile:
        text = infile.read()
    return text

def scrape_docx(file_name):
    doc = docx.Document(file_name)
 
    #pull in pararaphs
    fullText = []
    for para in doc.paragraphs:
        fullText.append(para.text)
 
    joinedText = '\n'.join(fullText)
 
    if len(doc.tables)>0:
      #pull in tables afterwards
      fullTable = ""
      for table in doc.tables:
          for row in table.rows:
              row_data = [cell.text for cell in row.cells]
              fullTable += ", ".join(row_data) + "\n"
      return (joinedText + '\n ' + fullTable, True)
    else:
      return (joinedText, False)

def scrape_xlsx(file_name):
  #pulls in a dictionary object of all tabs
  all_tabs_dict = pd.read_excel(file_name, na_values='', sheet_name=None)
 
  fullText = ""
  for df in list(all_tabs_dict.values()):
    for i in range(df.shape[0]):
        row_data = [str(df.iloc[i,j]) for j in range(df.shape[1])]
        fullText += ", ".join(row_data) + "\n"
 
  #remove multiple commas resulting from blank cells      
  fullText = re.sub(r',{2,}', ',', fullText)
  return fullText

def scrape_png(file_name):
    text = str(pytesseract.image_to_string(Image.open(file_name)))
    return text

def get_dataframe():    
    path ='C:/Users/HaoLuo/ACF-2023-0011/ACF-2023-0011' # set your own path
    files = os.listdir(path)
    os.chdir(path)
    set([x.split('.')[-1] for x in files])

    files_html = [x for x in files if x.endswith('.html') and x.startswith("ACF-")]
    html_dict_list=[]
    for file_path in files_html:
      #load html
      with open(path+"/"+file_path, 'r') as file:
        html_string=file.read()
        soup = BeautifulSoup(html_string, 'html.parser')
        print(file_path)
      #clean if public submission
      if soup.find_all(name='h1')[0].contents == ["PUBLIC SUBMISSION"]:
        html_dict = scrape_html(soup)

        #pull file paths for matching attachments
        attachment_paths = [x for x in files if html_dict['Document'] in x and not x.endswith(".html")]
        attachment_found = False
        html_dict["attachment_count"]=len(attachment_paths) #new line

        for attachment_path in attachment_paths:
            html_dict["attachment_name"] = attachment_path
            if attachment_path.endswith('.pdf'):
                html_dict["attachment_text"] = scrape_pdf(path+"/"+attachment_path)
                html_dict_list.append(html_dict.copy())
                attachment_found = True
            elif attachment_path.endswith('.docx'):
                html_dict["attachment_text"], html_dict["attachment_has_tables"] = scrape_docx(path+"/"+attachment_path) # new line
                html_dict["attachment_error"] = (len("attachment_text")==0)
                html_dict_list.append(html_dict.copy())
                attachment_found = True
            elif attachment_path.endswith('.rtf'):
                html_dict["attachment_text"] = scrape_rtf(path+"/"+attachment_path)
                html_dict_list.append(html_dict.copy())
                attachment_found = True
            elif attachment_path.endswith('.txt'):
                html_dict["attachment_text"] = scrape_txt(path+"/"+attachment_path)
                html_dict_list.append(html_dict.copy())
                attachment_found = True
            elif attachment_path.endswith('.png'):
                html_dict["attachment_text"] = scrape_png(path+"/"+attachment_path)
                html_dict_list.append(html_dict.copy())
                attachment_found = True
            elif attachment_path.endswith('.xlsx'):
                html_dict["attachment_text"] = scrape_xlsx(path+"/"+attachment_path)
                html_dict_list.append(html_dict.copy())
                attachment_found = True

        if attachment_found == False:
          #add row without attachment text
          html_dict_list.append(html_dict)

    df=pd.DataFrame(html_dict_list)
    
    for col_name in ["Comment","attachment_text"]:
      col_name_new = col_name+"_clean"

      #text cleaning
      df[col_name_new] = df[col_name].replace(to_replace=float('nan'),value='')

      df[col_name_new] = df[col_name_new].replace(to_replace='\xa0',value='')
      df[col_name_new] = df[col_name_new].str.replace(pat='§§',repl='bill number ')
      df[col_name_new] = df[col_name_new].str.replace(pat='§',repl='bill number ')
      df[col_name_new] = df[col_name_new].str.replace(pat='\u2009',repl='')

      df[col_name_new] = df[col_name_new].apply(lambda x: x.strip())
      df[col_name_new] = df[col_name_new].apply(lambda x: re.sub(r' +', ' ', x))
      df[col_name_new] = df[col_name_new].apply(lambda x: re.sub(r'\n +', '\n', x))
      df[col_name_new] = df[col_name_new].apply(lambda x: re.sub(r'\n+', '\n', x))
    return df

dff = get_dataframe()
congress = pd.read_excel("2024_jan_congress_members.xlsx")
congress_names = congress.Name.to_list()
 
dff["head_start_commenter"] = dff.Organization.str.contains(pat=r"Head Start.*?Association",case=False)
dff["congress_commenter"] = dff["Name"].apply(lambda x: x in congress_names)
dff.head()

dff.loc[~dff['Government Agency'].isna(), 'Government Agency'] = dff.loc[~dff['Government Agency'].isna(), 'Government Agency'].apply(lambda x: x[1:])

dff.loc[dff['Government Agency'].isin(["Senate HELP Committee, Chair Sanders",'U.S. House Committee on Education and the Workforce','Office of Congressman Adam B. Schiff']),'congress_commenter'] = True

dff[(dff['head_start_commenter']==1) | (dff['congress_commenter']==1)].to_excel('important_commenter.xlsx')

# single out importanta commenter and put html and attachment into a folder
import pandas as pd
import shutil
import os
import glob
dff1 = dff[(dff['head_start_commenter']==1) | (dff['congress_commenter']==1)]
file_source = 'C:\\Users\\HaoLuo\\ACF-2023-0011\\ACF-2023-0011'
file_destination = 'C:\\Users\\HaoLuo\\ACF-2023-0011\\important_commenter\\'
shutil.copy2('important_commenter.xlsx', file_destination)

os.chdir(file_source)
directory = glob.glob('*')
htm = [x+'.html' for x in dff1['Document']]

for file in directory:
    if file in dff1['attachment_name'].tolist():
        shutil.copy2(file_source+'\\'+file, file_destination)
    if file in htm:
        shutil.copy2(file_source+'\\'+file, file_destination)    

dff['Document']=np.where((dff['attachment_name'].apply(lambda x: str(x).split('.')[0])!='nan'),
dff['attachment_name'].apply(lambda x: str(x).split('.')[0]),
dff['Document'])

dff.to_pickle('2023_scrape.pkl')
print('Total Time consumed: ', time.time()-st)
